Moved transpose, tril and triu to new linalg submodule
