+++ heat/core/operations.py
@@ -1,5 +1,4 @@
 import builtins
-import itertools
 import numpy as np
 import torch
 import warnings
@@ -17,10 +16,7 @@ __all__ = [
     'allclose',
     'any',
     'argmax',
-    'argmin',
-    'transpose',
-    'tril',
-    'triu'
+    'argmin'
 ]
 
 
@@ -363,193 +359,6 @@ def argmin(x, axis=None, out=None, **kwargs):
         out._Tensor__array.storage().copy_(reduced_result._Tensor__array.storage())
 
     return reduced_result
-
-
-def transpose(a, axes=None):
-    """
-    Permute the dimensions of an array.
-
-    Parameters
-    ----------
-    a : array_like
-        Input array.
-    axes : None or list of ints, optional
-        By default, reverse the dimensions, otherwise permute the axes according to the values given.
-
-    Returns
-    -------
-    p : ht.Tensor
-        a with its axes permuted.
-    """
-    # type check the input tensor
-    if not isinstance(a, tensor.Tensor):
-        raise TypeError('a must be of type ht.Tensor, but was {}'.format(type(a)))
-
-    # set default value for axes permutations
-    dimensions = len(a.shape)
-    if axes is None:
-        axes = tuple(reversed(range(dimensions)))
-    # if given, sanitize the input
-    else:
-        try:
-            # convert to a list to allow index access
-            axes = list(axes)
-        except TypeError:
-            raise ValueError('axes must be an iterable containing ints')
-
-        if len(axes) != dimensions:
-            raise ValueError('axes do not match tensor shape')
-        for index, axis in enumerate(axes):
-            if not isinstance(axis, int):
-                raise TypeError('axis must be an integer, but was {}'.format(type(axis)))
-            elif axis < 0:
-                axes[index] = axis + dimensions
-
-    # infer the new split axis, it is the position of the split axis within the new axes permutation
-    try:
-        transposed_split = axes.index(a.split) if a.split is not None else None
-    except ValueError:
-        raise ValueError('axes do not match tensor shape')
-
-    # try to rearrange the tensor and return a new transposed variant
-    try:
-        transposed_data = a._Tensor__array.permute(*axes)
-        transposed_shape = tuple(a.shape[axis] for axis in axes)
-
-        return tensor.Tensor(transposed_data, transposed_shape, a.dtype, transposed_split, a.device, a.comm)
-    # if not possible re- raise any torch exception as ValueError
-    except RuntimeError as exception:
-        raise ValueError(str(exception))
-
-
-# statically allocated index slices for non-iterable dimensions in triangular operations
-__index_base = (slice(None), slice(None),)
-
-
-def __tri_op(m, k, op):
-    """
-    Generic implementation of triangle operations on tensors. It takes care of input sanitation and non-standard
-    broadcast behavior of the 2D triangle-operators.
-
-    Parameters
-    ----------
-    m : ht.Tensor
-        Input tensor for which to compute the triangle operator.
-    k : int, optional
-        Diagonal above which to apply the triangle operator, k<0 is below and k>0 is above.
-    op : callable
-        Implementation of the triangle operator.
-
-    Returns
-    -------
-    triangle_tensor : ht.Tensor
-        Tensor with the applied triangle operation
-
-    Raises
-    ------
-    TypeError
-        If the input is not a tensor or the diagonal offset cannot be converted to an integral value.
-    """
-    if not isinstance(m, tensor.Tensor):
-        raise TypeError('Expected m to be a tensor but was {}'.format(type(m)))
-
-    try:
-        k = int(k)
-    except ValueError:
-        raise TypeError('Expected k to be integral, but was {}'.format(type(k)))
-
-    # chunk the global shape of the tensor to obtain the offset compared to the other ranks
-    offset, _, _ = m.comm.chunk(m.shape, m.split)
-    dimensions = len(m.shape)
-
-    # manually repeat the input for vectors
-    if dimensions == 1:
-        triangle = m._Tensor__array.expand(m.shape[0], -1)
-        if torch.numel(triangle > 0):
-            triangle = op(triangle, k - offset)
-
-        return tensor.Tensor(
-            triangle,
-            (m.shape[0], m.shape[0],),
-            m.dtype,
-            None if m.split is None else 1,
-            m.device,
-            m.comm
-        )
-
-    original = m._Tensor__array
-    output = original.clone()
-
-    # modify k to account for tensor splits
-    if m.split is not None:
-        if m.split + 1 == dimensions - 1:
-            k += offset
-        elif m.split == dimensions - 1:
-            k -= offset
-
-    # in case of two dimensions we can just forward the call to the callable
-    if dimensions == 2:
-        if torch.numel(original) > 0:
-            op(original, k, out=output)
-    # more than two dimensions: iterate over all but the last two to realize 2D broadcasting
-    else:
-        ranges = [range(elements) for elements in m.lshape[:-2]]
-        for partial_index in itertools.product(*ranges):
-            index = partial_index + __index_base
-            op(original[index], k, out=output[index])
-
-    return tensor.Tensor(output, m.shape, m.dtype, m.split, m.device, m.comm)
-
-
-def tril(m, k=0):
-    """
-    Returns the lower triangular part of the tensor, the other elements of the result tensor are set to 0.
-
-    The lower triangular part of the tensor is defined as the elements on and below the diagonal.
-
-    The argument k controls which diagonal to consider. If k=0, all elements on and below the main diagonal are
-    retained. A positive value includes just as many diagonals above the main diagonal, and similarly a negative
-    value excludes just as many diagonals below the main diagonal.
-
-    Parameters
-    ----------
-    m : ht.Tensor
-        Input tensor for which to compute the lower triangle.
-    k : int, optional
-        Diagonal above which to zero elements. k=0 (default) is the main diagonal, k<0 is below and k>0 is above.
-
-    Returns
-    -------
-    lower_triangle : ht.Tensor
-        Lower triangle of the input tensor.
-    """
-    return __tri_op(m, k, torch.tril)
-
-
-def triu(m, k=0):
-    """
-    Returns the upper triangular part of the tensor, the other elements of the result tensor are set to 0.
-
-    The upper triangular part of the tensor is defined as the elements on and below the diagonal.
-
-    The argument k controls which diagonal to consider. If k=0, all elements on and below the main diagonal are
-    retained. A positive value includes just as many diagonals above the main diagonal, and similarly a negative
-    value excludes just as many diagonals below the main diagonal.
-
-    Parameters
-    ----------
-    m : ht.Tensor
-        Input tensor for which to compute the upper triangle.
-    k : int, optional
-        Diagonal above which to zero elements. k=0 (default) is the main diagonal, k<0 is below and k>0 is above.
-
-    Returns
-    -------
-    upper_triangle : ht.Tensor
-        Upper triangle of the input tensor.
-    """
-    return __tri_op(m, k, torch.triu)
-
 
 def __local_operation(operation, x, out):
     """
